Create routes.js and app.js
